apps/model-api/pyproject.toml
[tool.poetry]
name = "model-api"
version = "0.1.0"
description = "CITYFARM model API service"
authors = ["CITYFARM Team"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
flask = "^3.1.2"
python-dotenv = "^1.1.1"
gunicorn = "^23.0.0"
asgiref = "^3.10.0"
uvicorn = { version = "^0.38.0", extras = ["standard"] }

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
